requirements.txt
python-dotenv==1.0.0
requests==2.31.0
requests-toolbelt==1.0.0
pandas==2.1.3
knackpy==1.0.0
plaid-python==18.1.0
reportlab==4.0.7
Pillow==10.1.0
boto3==1.34.5
playwright==1.40.0
undetected-chromedriver==3.5.3
cryptography==41.0.7
tenacity==8.2.3
aiohttp==3.9.1
